Improved TWITransmitData function
Transmit data function no longer hangs whilst it is waiting for TWI
hardware to become ready. Instead it will return a value of 2 if
hardware was not ready. The application should check for this return
code and if a 2 was returned then the application should re-attempt the
transmission. This allows other operations to be performed whilst
waiting for the bus to become ready and does not lock the device up in a
situation such as disconnected lines.

This functionality is also exposed in the TWIReadData function (since it
must make a call to TWITransmitData to address the slave)